♻️ Reduce the number of lines in the src/options/sass.ts file
